refactor(browser): keep contracts inside the plugin package
