+++ packages/sdk/test/browser-plugin.test.ts
@@ -5,7 +5,7 @@ import { AppNodeBuilder } from "@opencode-ai/core/effect/app-node-builder"
 import { Location } from "@opencode-ai/core/location"
 import { Tool } from "@opencode-ai/core/tool"
 import plugin from "@opencode-ai/plugin-browser"
-import { Browser } from "@opencode-ai/schema/browser"
+import { Browser } from "@opencode-ai/plugin-browser/rpc"
 import { Agent, Rpc } from "@opencode-ai/plugin/effect"
 import type { Info } from "@opencode-ai/schema/tool"
 import { AbsolutePath, OpenCode, SessionMessage } from "@opencode-ai/sdk/effect"
